Feat: Add pw, loginId logic in finding pw and loginId
