+++ src/main/java/com/hangangnow/mainserver/config/redis/RedisConfig.java
@@ -1,14 +1,16 @@
-package com.hangangnow.mainserver.config;
+package com.hangangnow.mainserver.config.redis;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.PropertySource;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
+@PropertySource("classpath:/application-secret.properties")
 public class RedisConfig {
 
     @Value("${spring.redis.host}")
