+++ src/main/java/com/hangangnow/mainserver/config/security/SecurityConfig.java
@@ -1,9 +1,11 @@
-package com.hangangnow.mainserver.config;
+package com.hangangnow.mainserver.config.security;
 
 import com.hangangnow.mainserver.config.jwt.JwtAccessDeniedHandler;
 import com.hangangnow.mainserver.config.jwt.JwtAuthenticationEntryPoint;
 import com.hangangnow.mainserver.config.jwt.JwtSecurityConfig;
 import com.hangangnow.mainserver.config.jwt.TokenProvider;
+import com.hangangnow.mainserver.config.security.KakaoAuthenticationProvider;
+import com.hangangnow.mainserver.config.security.MemberAuthenticationProvider;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
