+++ src/main/java/com/hangangnow/mainserver/service/AuthService.java
@@ -1,30 +1,22 @@
 package com.hangangnow.mainserver.service;
 
-import com.hangangnow.mainserver.config.MemberAuthenticationProvider;
+import com.hangangnow.mainserver.config.security.MemberAuthenticationProvider;
 
-import com.hangangnow.mainserver.config.RedisUtil;
+import com.hangangnow.mainserver.config.redis.RedisUtil;
 import com.hangangnow.mainserver.domain.common.ResponseDto;
-import com.hangangnow.mainserver.domain.common.GenericResponseDto;
 import com.hangangnow.mainserver.domain.member.Member;
 import com.hangangnow.mainserver.domain.member.MemberProvider;
-import com.hangangnow.mainserver.domain.member.RefreshToken;
 import com.hangangnow.mainserver.domain.member.dto.*;
 import com.hangangnow.mainserver.config.jwt.TokenProvider;
 import com.hangangnow.mainserver.repository.MemberRepository;
-import com.hangangnow.mainserver.repository.RefreshTokenRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
 
 @Slf4j
 @Service
@@ -149,6 +141,10 @@ public class AuthService {
         Member member = memberRepository.findByEmail(passwordRequestDto.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("이메일이 존재하지 않습니다."));
 
+        if (member.getMemberProvider() == MemberProvider.KAKAO){
+            throw new RuntimeException("카카오 로그인으로 가입한 회원입니다. 카카오 로그인을 이용하세요");
+        }
+
         String password1 = passwordRequestDto.getPassword1();
         String password2 = passwordRequestDto.getPassword2();
 
@@ -168,6 +164,10 @@ public class AuthService {
     public ResponseDto findLoginIdByEmail(LoginIdRequestDto loginIdRequestDto) {
         Member findMember = memberRepository.findByEmail(loginIdRequestDto.getEmail())
                 .orElseThrow(() -> new IllegalArgumentException("해당 이메일을 가진 아이디가 존재하지 않습니다"));
+
+        if (findMember.getMemberProvider() == MemberProvider.KAKAO){
+            throw new RuntimeException("카카오 로그인으로 가입한 회원입니다. 카카오 로그인을 이용하세요");
+        }
 
         mailService.authLoginId(loginIdRequestDto.getEmail(), loginIdRequestDto.getName(), findMember.getLoginId());
         return new ResponseDto(loginIdRequestDto.getName() + "님 이메일로 아이디가 전송되었습니다.");
