+++ src/main/resources/application.properties
@@ -23,9 +23,5 @@ spring.mvc.pathmatch.matching-strategy=ant_path_matcher
 spring.mvc.throw-exception-if-no-handler-found=true
 spring.web.resources.add-mappings=false
 
-
-spring.redis.host=127.0.0.1
-spring.redis.port=6379
-
 logging.level.org.hibernate.SQL= debug
